girepository: Use size_t for size-related arguments and return types
We used to use unsigned values, while they should be big enough to old
the data we're handling here, it's cleaner and clearer if we use size_t
as type for such values, as it makes straight forward to understand what
a value should contain. It also makes these values more future proof.

+++ girepository/gibaseinfo.h
@@ -96,6 +96,6 @@ GI_AVAILABLE_IN_ALL
 GIBaseInfo *           gi_info_new                   (GIInfoType  type,
                                                       GIBaseInfo *container,
                                                       GITypelib  *typelib,
-                                                      uint32_t    offset);
+                                                      size_t      offset);
 
 G_END_DECLS
